requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.12.0
aiohttp==3.9.3
python-dateutil==2.8.2
email-validator==2.1.0.post1
jinja2==3.1.3
markdown2==2.4.12
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 